# Adding `Box` component in TS to component-library