--- a/ui/components/component-library/box/index.ts
+++ b/ui/components/component-library/box/index.ts
@@ -1,0 +1,8 @@
+export { Box } from './box';
+export type {
+  PolymorphicRef,
+  PolymorphicComponentPropWithRef,
+  StyleUtilityProps,
+  BoxProps,
+  BoxComponent,
+} from './box.types';