--- a/ui/components/component-library/component-library-components.scss
+++ b/ui/components/component-library/component-library-components.scss
@@ -5,6 +5,7 @@
 * unintended overrides.
 **/
 // Atoms
+@import 'box/box';
 @import 'text/text';
 @import 'icon/icon';
 @import 'label/label';