… ui/components/component-library/box/box.stories.tsx
… ui/components/component-library/box/box.test.tsx
… ui/components/component-library/box/box.tsx